refactor: refactor files to atmoic architecture
