+++ frontend/src/components/molecules/Calender/Calender.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { DayPicker } from "react-day-picker";
 
 import { cn } from "@/lib/utils";
-import { buttonVariants } from "@/components/ui/Button";
+import { buttonVariants } from "@/components/atoms/Button/Button";
 
 export type CalendarProps = React.ComponentProps<typeof DayPicker>;
